Renamed distance function and fixed monster spawning
Monsters kept spawning inside the "safe" zone, leading to some unfait situations. This was because the distance function was calculating distance in a 3D space, so when the random spawn point function chose a point slightly in the air, the distance would technically come out to being outside the radius of the bubble, and so the monster would spawn and fall into the "safe" zone.  Especially during the boss fight, you don't really want that to happen. This change simply ignores the y position and calculates distance on the XZ plane, and thus should guarantee they spawn outside the "safe" zone.